Update patch set 1
Patch Set 1:

Hmm, I was under the impression these were only for WCNSS, are you actually experiencing bluetooth issues that these fix?

Its definitely going to error out, since we share init (d851 is Broadcom, not WCNSS) and I'd rather not de-commonize it for things like this.

On WCNSS, this binary/lib is responsible for setting the MAC address of the bluetooth adapter.

Patch-set: 1
Label: Code-Review=0